feat: keep sip-019 notifs

+++ migrations/1692034871242_token-metadata-notifications.ts
@@ -42,7 +42,7 @@ export function up(pgm: MigrationBuilder): void {
       notNull: true,
     },
     ttl: {
-      type: 'int',
+      type: 'numeric',
     },
   });
   pgm.createConstraint(
